Create MAVConnection abstraction to implement forwarding.
